requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alpaca-trade-api==3.2.0
astroid==3.3.8
attrs==25.1.0
autopep8==2.3.2
beautifulsoup4==4.12.3
build==1.2.2.post1
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
deprecation==2.1.0
dill==0.3.9
fonttools==4.55.6
frozendict==2.4.6
frozenlist==1.5.0
html5lib==1.1
idna==3.10
isort==5.13.2
kiwisolver==1.4.8
lazy-object-proxy==1.10.0
lxml==5.3.0
matplotlib==3.10.0
mccabe==0.7.0
mplfinance==0.12.10b0
msgpack==1.0.3
multidict==6.1.0
multitasking==0.0.11
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pandas-datareader==0.10.0
peewee==3.17.8
pillow==11.1.0
pip-tools==7.4.1
platformdirs==4.3.6
propcache==0.2.1
pycodestyle==2.12.1
pylint==3.3.3
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.1
requests==2.32.3
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
ta==0.11.0
tabulate==0.9.0
toml==0.10.2
tomlkit==0.13.2
typed_ast==1.5.5
tzdata==2025.1
urllib3==1.26.20
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
yfinance==0.2.52
